+++ JavaExercises/Lab3/P5/Triplet.java
@@ -3,5 +3,5 @@ package JavaExercises.Lab3.P5;
 public class Triplet {
     public int a;
     public int b;
-    public int c;
+    public double c;
 }
